Inject files into package instead of creating symlinks - symlinks are still not working properly on windows (it fails deleting them because they're actually folders) - IDE indexes the symlinked folders while they're there which slows things down - any failure, like a file being in use, causes all symlinks to be left behind
